+++ src/rsnapshot_docker_compose_backup/config/default_config.py
@@ -3,12 +3,11 @@ from importlib import resources
 import os
 from pathlib import Path
 import re
-import sys
 from typing import Optional
 
 from rsnapshot_docker_compose_backup import global_values
 from rsnapshot_docker_compose_backup.config.abstract_config import AbstractConfig
-from rsnapshot_docker_compose_backup.utils import CaseInsensitiveRe
+from rsnapshot_docker_compose_backup.utils.utils import CaseInsensitiveRe
 
 
 class DefaultConfig(AbstractConfig):
